pat-discord-lnbits-bot/backend/requirements.txt
discord.py==2.3.2
aiohttp==3.9.1
websockets==12.0
requests==2.31.0
qrcode==7.4.2
Pillow==10.1.0
flask==3.0.0
flask-cors==4.0.0